dailyblog/requirements.txt
beautifulsoup4==4.6.0
celery==3.1.25
chardet==2.3.0
Django==1.9.6
django-celery==3.2.1
django-ckeditor==5.3.0
django-filter==0.13.0
django-filters==0.2.1
djangorestframework==3.3.3
Fabric==1.11.1
gevent==1.1.1
gunicorn==19.6.0
html5lib==0.999
httplib2==0.8
itsdangerous==0.24
jobspider==2.3.3
lxml==3.7.1
Mako==1.0.4
Markdown==2.6.6
MarkupSafe==0.23
MySQL-python==1.2.5
paramiko==1.17.0
Pillow==2.3.0
python-memcached==1.58
qcloud-cos-v4==0.0.22
qrcode==5.3
requests==2.18.3
six==1.5.2
supervisor==3.3.0
urllib3==1.22
weibo==0.2.2
